requirements.txt
customtkinter
rawpy
opencv-python-headless
pillow
piexif
numpy